move some shared functions to /utils/
same as what was done in #224 with char2num, gtpchar2num, and num2char
