+++ game.js
@@ -1,8 +1,8 @@
 // vim: tw=120 softtabstop=4 shiftwidth=4
 
+const { decodeMoves }    = require("./utils/decodeMoves");
+const { move2gtpvertex } = require("./utils/move2gtpvertex");
 const Bot = require('./bot').Bot;
-const decodeMoves = require('./bot').decodeMoves;
-const move2gtpvertex = require('./bot').move2gtpvertex;
 const console = require('./console').console;
 const config = require('./config');
